refactor(P5-002/P5-003): move WorkflowEffectivenessSurface to Workflows module
Relocates components/execution/WorkflowEffectivenessSurface.tsx to
components/Workflows/WorkflowEffectivenessSurface.tsx so workflow
diagnostics are owned by the Workflows module rather than execution/.

Updates all consumers to import from the new path:
- AnalyticsDashboard.tsx (workflow_intelligence tab — still renders the
  surface; Analytics discovers it, Workflows owns it)
- FeatureExecutionWorkbench.tsx (embedded effectiveness view)

The ArtifactReferenceModal import inside the surface is updated to cross
into ../execution/ since that helper remains execution-owned.
No route, prop, or render behavior changes. Test baseline unchanged
(20 pre-existing failures, 0 new).
